--- exercicios/para-casa/entregas/exemplo-nome-sobrenome/resolucao.test.js
@@ -1,17 +1,0 @@
-const request = require('supertest')
-const ApiUrl = "https://servicodados.ibge.gov.br/api/v1/localidades"
-
-describe('GET /estados', () => {
-    
-    test('retorna 200 quando buscar pelo município de Ananindeua', () => {
-        const estadoPara = () => expect.arrayContaining([expect.objectContaining({ "nome": "Pará" })])
-        const municipioAnanindeua = estadoPara([expect.objectContaining({ "nome": "Ananindeua" })])
-
-        request(ApiUrl)
-            .get('/estados')
-            .expect(200)
-            .then(response => {
-                expect(response.body).toEqual(municipioAnanindeua)
-            })
-    })
-})
